requirements.txt
Flask==2.0.1
Flask-Admin==1.6.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-RESTful==0.3.9
python-crontab==2.5.1
psycopg2-binary==2.9.1
Werkzeug==2.0.1
python-dotenv==0.19.0
Flask-Migrate==3.1.0 